requirements.txt
anyio==3.6.2
astroid==2.15.2
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
delocate==0.10.4
dill==0.3.6
flake8==6.0.0
Flask==2.2.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
pycodestyle==2.10.0
pyflakes==3.0.1
pylint==2.17.2
python-dotenv==1.0.0
python-telegram-bot==20.2
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
tomli==2.0.1
tomlkit==0.11.7
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
